Implement client read threading
