+++ threads/Thread.cpp
@@ -1,5 +1,4 @@
 #include "Thread.hpp"
-#include "SocketThread.hpp"
 #include <pthread.h>
 #include <stdlib.h>
 #include <string.h>
